+++ install.sh
@@ -4,11 +4,8 @@ set -eu
 # https://github.com/objectbox/objectbox-c/releases
 # It's important that the generated dart bindings and the c-api library version match. Dart won't error on C function
 # signature mismatch, leading to obscure memory bugs.
-# After upgrading to a different C-API version:
-#   * update lib/src/bindings/objectbox.h
-#   * execute pub run ffigen
-#   * have a look at the changed files to see if some call sites need to be updated
-cLibVersion=0.14.0
+# For how to upgrade the version see dev-doc/updating-c-library.md
+cLibVersion=0.15.2
 os=$(uname)
 cLibArgs="$*"