+++ objectbox/lib/src/native/box.dart
@@ -16,7 +16,6 @@ import 'bindings/bindings.dart';
 import 'bindings/flatbuffers.dart';
 import 'bindings/helpers.dart';
 import 'query/query.dart';
-import 'transaction.dart';
 
 /// Box put (write) mode.
 enum PutMode {
@@ -360,8 +359,9 @@ class Box<T> {
 
   void _putToManyRelFields(T object, PutMode mode, Transaction tx) {
     _entity.toManyRelations(object).forEach((RelInfo info, ToMany rel) {
+      // Always set relation info so ToMany applyToDb can be used after initial put
+      InternalToManyAccess.setRelInfo(rel, _store, info, this);
       if (InternalToManyAccess.hasPendingDbChanges(rel)) {
-        InternalToManyAccess.setRelInfo(rel, _store, info, this);
         rel.applyToDb(mode: mode, tx: tx);
       }
     });
@@ -380,6 +380,8 @@ int _getOBXPutMode(PutMode mode) {
   return __getOBXPutMode(mode);
 }
 
+/// Something like an "async box": keeps an OBX_async pointer and offers
+/// functions to handle async messages (e.g. via ReceivePort).
 class _AsyncBoxHelper {
   final Pointer<OBX_async> _cAsync;
 
@@ -387,6 +389,10 @@ class _AsyncBoxHelper {
     initializeDartAPI();
   }
 
+  /// Put the given buffer as an object asynchronously.
+  /// Internally, this depends on the dartc_async_put_object C function,
+  /// which call an obx_async_put function with a callback. The callback sends
+  /// an empty message if successful, or an error string.
   Future<int> put(int id, BuilderWithCBuffer fbb, PutMode mode) async {
     final port = ReceivePort();
     final newId = C.dartc_async_put_object(_cAsync, port.sendPort.nativePort,