+++ objectbox/test/box_test.dart
@@ -16,7 +16,7 @@ void main() {
   late Store store;
   late Box<TestEntity> box;
 
-  final simpleItems = () => ['One', 'Two', 'Three', 'Four', 'Five', 'Six']
+  simpleItems() => ['One', 'Two', 'Three', 'Four', 'Five', 'Six']
       .map((s) => TestEntity(tString: s))
       .toList();
 
@@ -26,7 +26,7 @@ void main() {
     store = env.store;
   });
 
-  tearDown(() => env.close());
+  tearDown(() => env.closeAndDelete());
 
   test('store box vending', () {
     final box1 = store.box<TestEntity>();
@@ -201,6 +201,45 @@ void main() {
         () => box.putMany([u1, again]),
         throwsA(predicate((UniqueViolationException e) =>
             e.toString().contains('same property value already exists'))));
+  });
+
+  test('.put() replaces duplicate values on a unique replace field on insert',
+      () {
+    // insert without conflict
+    box.putMany([
+      TestEntity.uniqueReplace(replaceLong: 1, tString: 'original-1'),
+      TestEntity.uniqueReplace(replaceLong: 2, tString: 'original-2')
+    ]);
+    expect(box.count(), equals(2));
+
+    // insert with conflict, deletes ID 1 and inserts ID 3
+    box.put(TestEntity.uniqueReplace(replaceLong: 1, tString: 'replacement-1'));
+    expect(box.count(), equals(2));
+    final replaced = box.get(3)!;
+    expect(replaced.replaceLong, equals(1));
+    expect(replaced.tString, equals('replacement-1'));
+  });
+
+  test('.put() replaces duplicate values on a unique replace field on update',
+      () {
+    // update without conflict
+    var first = TestEntity.uniqueReplace(replaceLong: 1, tString: 'first');
+    box.put(first);
+    first.replaceLong = 2;
+    box.put(first);
+    expect(box.count(), equals(1));
+    final updated = box.get(1)!;
+    expect(updated.replaceLong, equals(2));
+    expect(updated.tString, 'first');
+
+    // update with conflict, deletes ID 2 and keeps ID 1
+    box.put(TestEntity.uniqueReplace(replaceLong: 1, tString: 'second'));
+    first.replaceLong = 1;
+    box.put(first);
+    expect(box.count(), equals(1));
+    final updated2 = box.get(1)!;
+    expect(updated2.replaceLong, equals(1));
+    expect(updated2.tString, 'first');
   });
 
   test('.getAll retrieves all items', () {
@@ -441,7 +480,7 @@ void main() {
     bool contains = box.contains(id);
     expect(contains, equals(true));
     //check complementary
-    box.remove(id);
+    expect(box.remove(id), equals(true));
     contains = box.contains(id);
     expect(contains, equals(false));
   });
@@ -451,7 +490,7 @@ void main() {
     bool contains = box.containsMany(ids);
     expect(contains, equals(true));
     //check with one missing id
-    box.remove(ids[1]);
+    expect(box.remove(ids[1]), isTrue);
     contains = box.containsMany(ids);
     expect(contains, equals(false));
     //check complementary
@@ -463,12 +502,12 @@ void main() {
   test('.remove(id) works', () {
     final List<int> ids = box.putMany(simpleItems());
     //check if single id remove works
-    expect(box.remove(ids[1]), equals(true));
+    expect(box.remove(ids[1]), isTrue);
     expect(box.count(), equals(5));
     //check what happens if id already deleted -> throws OBJBOXEX 404
     bool success = box.remove(ids[1]);
     expect(box.count(), equals(5));
-    expect(success, equals(false));
+    expect(success, isFalse);
   });
 
   test('.remove() returns false on non-existent item', () {