+++ tool/publish.sh
@@ -1,14 +1,29 @@
 . "$(dirname "$0")"/common.sh
 
+# NOTE: This script requires version 3 of yq.
+# Download versions for other architectures from https://github.com/mikefarah/yq/releases/tag/3.4.1
+msys=false
+case "$( uname )" in                #(
+  MSYS* | MINGW* )  msys=true    ;; #(
+esac
+if [ $msys ]; then
+  YQCMD="${root}/tool/yq_windows_amd64.exe"
+else
+  YQCMD="${root}/tool/yq_linux_amd64"
+fi
+echo "Testing yq version..."
+$YQCMD -V
+echo "Testing yq version...DONE"
+
 # ======================= BEFORE publishing==================== #
 
 echo "Removing dependency_overrides from all pubspec.yaml files (backup at pubspec.yaml.original)"
 find "${root}" -type f -name "pubspec.yaml" \
   -exec echo "Processing {}" \; \
   -exec cp "{}" "{}.original" \; \
-  -exec yq delete -i "{}" dependency_overrides \;
+  -exec "$YQCMD" delete -i "{}" dependency_overrides \;
 
-# update links in the readme (see `git restore "${root}/objectbox/README.md"` below)
+# Update links in READMEs (restored by git restore commands below).
 "${root}/tool/pubdev-links.sh"
 
 # =========================== PUBLISH ======================== #
@@ -21,16 +36,16 @@ function publish() {
   pkg_dir="${root}/${1}"
   pubspec="${pkg_dir}/pubspec.yaml"
 
-  echo -e "You're about to publish directory \e[33m'${1}'\e[39m as package \e[33m$(yq read "${pubspec}" name) v$(yq read "${pubspec}" version)\e[39m"
+  echo -e "You're about to publish directory \e[33m'${1}'\e[39m as package \e[33m$($YQCMD read "${pubspec}" name) v$($YQCMD read "${pubspec}" version)\e[39m"
   echo -e "\e[31mWARNING: The same version can NOT be published twice!\e[39m"
-  read -p " Are you sure you want to publish to pub.dev? " yn
+  read -p " Publish to pub.dev [y/N]? " yn
   case $yn in
   [Yy]*)
     cd "${pkg_dir}" || exit 1
-    pub publish --force
+    dart pub publish --force
     ;;
   [Nn]*) ;;
-  *) echo "Please answer yes or no." ;;
+  *) echo "Not publishing this package." ;;
   esac
 }
 
@@ -48,3 +63,5 @@ find "${root}" -type f -name "pubspec.yaml" \
 
 echo "Restoring objectbox/README.md"
 git restore "${root}/objectbox/README.md"
+echo "Restoring objectbox/example/README.md"
+git restore "${root}/objectbox/example/README.md"